requirements.txt
streamlit>=1.29.0
tqdm>=4.66.0
python-dotenv>=1.0.0
PyYAML>=6.0
lxml>=4.9.3
pydantic>=2.5.0
pandas>=2.0.0
xlsxwriter>=3.1.0
matplotlib>=3.7.0
numpy>=1.24.0
pywin32>=302
# SQLite3 is included in Python standard library 